pyproject.toml
[tool.poetry]
name = "campaign_creator_v2"
version = "0.1.0"
description = "Campaign Creator for RPG's."
authors = ["Wouter Lansu <wfrlansu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
django-debug-toolbar = "^3.6.0"
Pillow = "^9.2.0"
django-model-utils = "^4.2.0"
whitenoise = "^6.2.0"
django-storages = "^1.13.1"
django-allauth = "^0.51.0"
crispy-bootstrap5 = "^0.6"
django-crispy-forms = "^1.14.0"
django-environ = "^0.9.0"
argon2-cffi = "^21.3.0"
rcssmin = "1.1.0"
django-compressor = "^4.1"
pytz = "^2022.2.1"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
python-slugify = "^6.1.2"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.2.1"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.5"
django-stubs = "^1.12.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "campaigncreator.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
filterwarnings = [
    "ignore::UserWarning",
]
